Move sendToLeaders to UpdateReqeust
