+++ solr/solrj/src/java/org/apache/solr/client/solrj/SolrRequest.java
@@ -237,16 +237,6 @@ public abstract class SolrRequest<T> implements Serializable {
   }
 
   /**
-   * Indicates if clients should make attempts to route this request to a shard leader, overriding
-   * typical client routing preferences for requests. Defaults to true.
-   *
-   * @see CloudSolrClient#isUpdatesToLeaders
-   */
-  public boolean shouldSendToLeaders() {
-    return true;
-  }
-
-  /**
    * Indicates which API version this request will make
    *
    * <p>Defaults implementation returns 'V1'.
