+++ solr/solrj/src/java/org/apache/solr/client/solrj/impl/CloudSolrClient.java
@@ -52,7 +52,6 @@ import org.apache.solr.client.solrj.SolrClient;
 import org.apache.solr.client.solrj.SolrRequest;
 import org.apache.solr.client.solrj.SolrRequest.SolrRequestType;
 import org.apache.solr.client.solrj.SolrServerException;
-import org.apache.solr.client.solrj.request.AbstractUpdateRequest;
 import org.apache.solr.client.solrj.request.RequestWriter;
 import org.apache.solr.client.solrj.request.UpdateRequest;
 import org.apache.solr.client.solrj.request.V2Request;
@@ -355,9 +354,8 @@ public abstract class CloudSolrClient extends SolrClient {
   }
 
   @SuppressWarnings({"unchecked"})
-  private NamedList<Object> directUpdate(AbstractUpdateRequest request, String collection)
+  private NamedList<Object> directUpdate(UpdateRequest request, String collection)
       throws SolrServerException {
-    UpdateRequest updateRequest = (UpdateRequest) request;
     SolrParams params = request.getParams();
     ModifiableSolrParams routableParams = new ModifiableSolrParams();
     ModifiableSolrParams nonRoutableParams = new ModifiableSolrParams();
@@ -411,9 +409,9 @@ public abstract class CloudSolrClient extends SolrClient {
     String routeField =
         (col.getRouter().getRouteField(col) == null) ? ID : col.getRouter().getRouteField(col);
     final Map<String, ? extends LBSolrClient.Req> routes =
-        createRoutes(updateRequest, routableParams, col, router, urlMap, routeField);
+        createRoutes(request, routableParams, col, router, urlMap, routeField);
     if (routes == null) {
-      if (directUpdatesToLeadersOnly && hasInfoToFindLeaders(updateRequest, routeField)) {
+      if (directUpdatesToLeadersOnly && hasInfoToFindLeaders(request, routeField)) {
         // we have info (documents with ids and/or ids to delete) with
         // which to find the leaders, but we could not find (all of) them
         throw new SolrException(
@@ -490,7 +488,7 @@ public abstract class CloudSolrClient extends SolrClient {
     }
 
     UpdateRequest nonRoutableRequest = null;
-    List<String> deleteQuery = updateRequest.getDeleteQuery();
+    List<String> deleteQuery = request.getDeleteQuery();
     if (deleteQuery != null && deleteQuery.size() > 0) {
       UpdateRequest deleteQueryRequest = new UpdateRequest();
       deleteQueryRequest.setDeleteQuery(deleteQuery);
@@ -1001,24 +999,29 @@ public abstract class CloudSolrClient extends SolrClient {
     boolean sendToLeaders = false;
 
     if (request.getRequestType() == SolrRequestType.UPDATE) {
-      sendToLeaders = request.shouldSendToLeaders() && this.isUpdatesToLeaders();
+      sendToLeaders = this.isUpdatesToLeaders();
 
-      // Check if we can do a "directUpdate" ...
       if (sendToLeaders && request instanceof UpdateRequest) {
-        if (inputCollections.size() > 1) {
-          throw new SolrException(
-              SolrException.ErrorCode.BAD_REQUEST,
-              "Update request must be sent to a single collection "
-                  + "or an alias: "
-                  + inputCollections);
-        }
-        String collection =
-            inputCollections.isEmpty()
-                ? null
-                : inputCollections.get(0); // getting first mimics HttpSolrCall
-        NamedList<Object> response = directUpdate((AbstractUpdateRequest) request, collection);
-        if (response != null) {
-          return response;
+        var updateRequest = (UpdateRequest) request;
+        sendToLeaders = sendToLeaders && updateRequest.isSendToLeaders();
+
+        // Check if we can do a "directUpdate" ...
+        if (sendToLeaders) {
+          if (inputCollections.size() > 1) {
+            throw new SolrException(
+                SolrException.ErrorCode.BAD_REQUEST,
+                "Update request must be sent to a single collection "
+                    + "or an alias: "
+                    + inputCollections);
+          }
+          String collection =
+              inputCollections.isEmpty()
+                  ? null
+                  : inputCollections.get(0); // getting first mimics HttpSolrCall
+          NamedList<Object> response = directUpdate(updateRequest, collection);
+          if (response != null) {
+            return response;
+          }
         }
       }
     }
@@ -1173,8 +1176,8 @@ public abstract class CloudSolrClient extends SolrClient {
   /**
    * If true, this client has been configured such that it will generally prefer to send {@link
    * SolrRequestType#UPDATE} requests to a shard leader, if and only if {@link
-   * SolrRequest#shouldSendToLeaders} is also true. If false, then this client has been configured
-   * to obey normal routing preferences when dealing with {@link SolrRequestType#UPDATE} requests.
+   * UpdateRequest#isSendToLeaders} is also true. If false, then this client has been configured to
+   * obey normal routing preferences when dealing with {@link SolrRequestType#UPDATE} requests.
    *
    * @see #isDirectUpdatesToLeadersOnly
    */
@@ -1186,7 +1189,7 @@ public abstract class CloudSolrClient extends SolrClient {
    * If true, this client has been configured such that "direct updates" will <em>only</em> be sent
    * to the current leader of the corresponding shard, and will not be retried with other replicas.
    * This method has no effect if {@link #isUpdatesToLeaders()} or {@link
-   * SolrRequest#shouldSendToLeaders} returns false.
+   * UpdateRequest#isSendToLeaders} returns false.
    *
    * <p>A "direct update" is any update that can be sent directly to a single shard, and does not
    * need to be broadcast to every shard. (Example: document updates or "delete by id" when using
